Fix classification store localized attr. error
For localized classification store attributes, the classificationKeyId is empty. The identification of the related attribute takes place based on the attribute name, which is fine.
However, for Pimcore X, the KeyConfig::getById() will cause an exception causing an error in the user interface, because the classificationKeyId is empty in that case.
To avoid the exception, an additional condition has been added.